refactor: move friend request logic to index.tsx and remove requests.tsx
