+++ src/components/friendRequestModal.tsx
@@ -1,27 +1,24 @@
-import { View, Text, Pressable, Animated } from "react-native";
+import { View, Text, Pressable, Animated, Alert } from "react-native";
 import { Modalize } from "react-native-modalize";
 import { useRef, useEffect } from "react";
 import { ScrollView } from "react-native-gesture-handler";
 import { FriendRequest } from "@/types";
 import FloatingAvatar from "@/components/floatingAvatar";
-import { BlurView } from 'expo-blur';;
+import { BlurView } from 'expo-blur';
+import AsyncStorage from "@react-native-async-storage/async-storage";
 
 type Props = {
   visible: boolean;
   request: FriendRequest | null;
   onClose: () => void;
-  onAccept: () => void;
-  onReject: () => void;
-  scaleAnim: Animated.Value;
+  onHandled: () => void;
 };
 
 export default function FriendRequestModal({ 
-    visible, 
-    request, 
-    onClose, 
-    onAccept,
-    onReject,
-    scaleAnim 
+  visible, 
+  request, 
+  onClose, 
+  onHandled
 }: Props) {
   const modalRef = useRef<Modalize>(null);
 
@@ -35,6 +32,41 @@ export default function FriendRequestModal({
 
   // Only render if request is present
   if (!request) return null;
+
+  const handleFriendRequest = async (
+      type: 'accept' | 'reject'
+  ) => {
+      const endpoint = type === 'accept' ? '/accept-request' : '/reject-request';
+
+      try {
+          const token = await AsyncStorage.getItem("access_token");
+          const res = await fetch(`${process.env.EXPO_PUBLIC_API_URL}/friends${endpoint}`, {
+              method: 'PATCH',
+              headers: { 
+                  'Content-Type': 'application/json', 
+                  Authorization: `Bearer ${ token }`
+              },
+              body: JSON.stringify({ 
+                id: request.id, 
+                uid1: request.uid1, 
+                uid2: request.uid2 
+              }),
+          });
+
+          if (!res.ok) {
+              const { error } = await res.json();
+              console.error("Error fetching requests:", error);
+              Alert.alert("Error", error || "Something went wrong.");
+              return;
+          }
+
+          onClose();
+          onHandled();
+      } catch (error) {
+          console.error('Error', error);
+          Alert.alert("Error", "Unable to process request.");
+      }
+  }
 
   return (
     <Modalize
@@ -63,12 +95,12 @@ export default function FriendRequestModal({
           </View>
 
           <View className="flex-row gap-4 mb-4">
-            <Pressable onPress={onAccept} className="border border-white px-4 py-2 rounded-lg items-center">
+            <Pressable onPress={() => handleFriendRequest("accept")} className="border border-white px-4 py-2 rounded-lg items-center">
               <Text className="text-2xl">✅</Text>
               <Text className="text-white font-medium mt-1">Accept</Text>
             </Pressable>
 
-            <Pressable onPress={onReject} className="border border-white px-4 py-2 rounded-lg items-center">
+            <Pressable onPress={() => handleFriendRequest("reject")} className="border border-white px-4 py-2 rounded-lg items-center">
               <Text className="text-2xl">❌</Text>
               <Text className="text-white font-medium mt-1">Decline</Text>
             </Pressable>
